+++ CHANGELOG.md
@@ -10,6 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * `hbf` FIRs, symmetric FIRs, half band filters, HBF decimators and interpolators
+* `iir::PidBuilder` a builder for PID coefficients
+* `iir::Biquad::{hold, proportional, identity}`
+
+### Removed
+
+* `iir::Vec5` type alias has been removed.
+
+### Changed
+
+* `iir`: The biquad IIR filter API has been reworked. `IIR -> Biquad` renamed.
 
 ## [0.10.0](https://github.com/quartiq/idsp/compare/v0.9.2..v0.10.0) - 2023-07-20